Update and Invoice Formating Done

+++ POSApp/POSApp/Models/CustomerInfo.cs
@@ -7,7 +7,7 @@ namespace POSApp.Models
     public class CustomerInfo
     {
         public string CustomerId { get; set; }
-        public decimal VoucherLimit { get; set; }
+        public double VoucherLimit { get; set; }
         public int FamilySize { get; set; }
 
     }
